requirements.txt
antlr4-python3-runtime==4.9.3
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
contourpy==1.0.6
cycler==0.11.0
docker-pycreds==0.4.0
docopt==0.6.2
exceptiongroup==1.0.1
filelock==3.8.0
flake8==5.0.4
fonttools==4.38.0
gitdb==4.0.9
GitPython==3.1.29
huggingface-hub==0.10.1
hydra-core==1.2.0
idna==3.4
importlib-resources==5.10.0
iniconfig==1.1.1
joblib==1.2.0
kiwisolver==1.4.4
matplotlib==3.6.1
mccabe==0.7.0
numpy==1.23.4
omegaconf==2.2.3
packaging==21.3
pandas==1.5.0
pathtools==0.1.2
Pillow==9.3.0
pipreqs==0.4.11
pluggy==1.0.0
promise==2.3
protobuf==4.21.9
psutil==5.9.4
pycodestyle==2.9.1
pyflakes==2.5.0
pyparsing==3.0.9
pytest==7.2.0
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
scikit-learn==1.1.3
scipy==1.9.3
sentry-sdk==1.10.1
setproctitle==1.3.2
shortuuid==1.0.10
six==1.16.0
smmap==5.0.0
threadpoolctl==3.1.0
tokenizers==0.13.2
tomli==2.0.1
torch==1.12.1
tqdm==4.64.1
transformers==4.23.1
typing_extensions==4.4.0
urllib3==1.26.12
wandb==0.13.5
yarg==0.1.9
zipp==3.10.0
